+++ wiki/Changelog.md
@@ -1,7 +1,10 @@
   * Changed non-graphic display substitute for SIXEL image to space to reduce flickering (~#587).
+  * Changed non-graphic clipboard substitute for SIXEL image to configurable string, default space (#587).
+
   * Support for "combining double" characters that cover two base characters (~#553).
-  * New command line option --dir to change the initial working directory (#558, #58).
   * Expanding characters as appropriate for wide display (#123, /#570).
+
+  * New command line option --dir to change the initial working directory (#558, #58).
 
 ### 2.6.1 (18 Sep 2016) ###